Model changes in ASN
