+++ src/app/core/models/advance-shipping-notification.ts
@@ -1,7 +1,3 @@
-import { Plants } from './plants';
-import { Products } from './products';
-import { StorageLocations } from './storage-location';
-import { Units } from './units';
 
 export interface AdvancedShipmentNotificationVM{
   ASNId?:number;
@@ -41,4 +37,52 @@ export interface AdvancedShipmentNotificationDetVM{
   DeliveryQty?: number;
   OpenGRQty: number,
   DeliveryDate: Date,
+}
+
+export interface ASNDetailsLine{
+  Id?: number,
+  LineId : number,
+  POHeaderId: number,
+  ProductId: number,
+  ProductCode: string,
+  ProductDescription: string,
+  ProductGroup: string,
+  POQty: number,
+  Qty: number,
+  OpenGRQty?: number,
+  DeliveryDate: Date,
+  UnitId: number,
+  UnitName: string,
+  UnitDescription: string,
+  PlantId: number,
+  PlantCode: string,
+  PlantDescription: string,
+  StorageLocationId: number,
+  LocationCode: string,
+  LocationDescription: string,
+  NetPrice?:number;
+  TotalNetPrice?:number;
+  Currency?:string;
+  HSNCode?:string;
+  GST?:number;
+  IGST?:number;
+  SGST?:number;
+  CGST?:number;
+  TaxAmount?:number;
+  TotalAmount?:number;
+  StockType?:string;
+  PRDetId?: number;
+  RFQDetId?: number;
+  OADetId?: number;
+  IsReturnItem?: boolean;
+  IsFreeOfCharge?: boolean;
+  IsActive: boolean,
+  CreatedBy: number,
+  CreatedOn: Date,
+  UpdatedBy: number,
+  UpdatedOn: Date,
+  IsDeleted: true,
+  DeletedOn: Date,
+  Extra1: string,
+  Extra2: string,
 }
